requirement.txt
ultralytics==8.3.228
roboflow==1.2.11
matplotlib==3.10.7
pandas==2.3.3
seaborn==0.13.2
Pillow==12.0.0
opencv-python==4.12.0.88
numpy==2.3.5
albumentations==2.0.8